requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Model Interpretability and Explainability
shap>=0.42.0
matplotlib>=3.5.0
seaborn>=0.11.0
captum>=0.6.0

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0

# Caching
cachetools>=5.0.0

# Reinforcement Learning
stable-baselines3>=1.6.0
gym>=0.21.0

# Distributed Computing
ray[default]>=2.0.0
ray[tune]>=2.0.0

# Data Processing
pydantic>=1.8.0
PyYAML>=6.0
python-dotenv>=0.19.0
email-validator>=2.1.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=4.0.0

# Time Series Database
influxdb-client>=1.36.0
asyncpg>=0.28.0  # For TimescaleDB async support

# API Framework
fastapi>=0.70.0
uvicorn>=0.15.0
python-multipart>=0.0.5
websockets>=10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP Client
httpx>=0.24.0
requests>=2.25.0

# Monitoring
psutil>=5.8.0
prometheus-client>=0.12.0

# Cloud integrations
boto3>=1.28.0

# Development
pytest>=6.2.0
pytest-asyncio>=0.18.0
black>=21.0.0
flake8>=4.0.0
mypy>=0.910

# Exchange APIs (placeholder - actual packages may vary)
# robinhood-api>=1.0.0  # Placeholder
# oandapyV20>=0.6.0
# coinbase-pro>=1.1.0
